requirements.txt
urllib3
python-telegram-bot